Finish serialization implementation.

+++ ftss-tests/ClearTests.cs
@@ -17,9 +17,9 @@ namespace ftss_tests
             FastTernaryStringSet test = [];
 
             // Act & Assert
-            Assert.AreEqual(0, test.Size, "Test A");
+            Assert.AreEqual((uint)0, test.Size, "Test A");
             test.Clear();
-            Assert.AreEqual(0, test.Size, "Test B");
+            Assert.AreEqual((uint)0, test.Size, "Test B");
         }
 
         [TestMethod]
@@ -30,9 +30,9 @@ namespace ftss_tests
 
             // Act & Assert
             test.AddAll(["chicken", "duck", "whale",]);
-            Assert.AreEqual(3, test.Size, "Test A");
+            Assert.AreEqual((uint)3, test.Size, "Test A");
             test.Clear();
-            Assert.AreEqual(0, test.Size, "Test B");
+            Assert.AreEqual((uint)0, test.Size, "Test B");
         }
 
         [TestMethod]
@@ -44,9 +44,9 @@ namespace ftss_tests
             test.Add(string.Empty);
 
             // Act & Assert
-            Assert.AreEqual(2, test.Size, "Test A");
+            Assert.AreEqual((uint)2, test.Size, "Test A");
             test.Clear();
-            Assert.AreEqual(0, test.Size, "Test B");
+            Assert.AreEqual((uint)0, test.Size, "Test B");
         }
     }
 }
